crud message chat

+++ studdyroom/base/views.py
@@ -1,12 +1,16 @@
+from email import message
 import queue
 import re
 from django.shortcuts import render, redirect
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 from django.contrib import messages
+
 from django.http import HttpResponse
-from .models import Room, Topic
+from .models import Room, Topic, Message
 from .forms import RoomForm
 # Create your views here.
 
@@ -20,10 +24,14 @@ rooms = [
 ]
 
 def loginPage(request):
-    context = {}
+    
+    page='login'
+    context = {'page': page}
+    if request.user.is_authenticated:
+        return redirect('home')
 
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
         try:
             user = User.objects.get(username=username)
@@ -36,6 +44,29 @@ def loginPage(request):
             return redirect('home')
         else:
             messages.error(request, 'user name or password is incorrect')
+    return render(request, 'base/login_register.html', context)
+
+
+def LogoutUser(request):
+    
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    page='Register'
+    
+    form = UserCreationForm()
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Something went wrong')
+    context = {'page': page, 'form': form}
     return render(request, 'base/login_register.html', context)
 
 def home(request):
@@ -52,11 +83,28 @@ def home(request):
 
 def room(request, pk):
     context = {}
-    for room in rooms:
-        if room['id'] == int(pk):
-            context = {'room': room}
+    room = Room.objects.get(id=pk)
+    participants = room.participants.all()
+    room_messages = room.message_set.all().order_by('-created')
+    if request.method == 'POST':
+        message = request.POST.get('body')
+        if message:
+            Message.objects.create(
+                user=request.user,
+                room=room,
+                body=message
+            )
+        room.participants.add(request.user)
+        return redirect('room', pk=room.id)
+        
+    context = {
+        'room': room,
+        'room_messages': room_messages,
+        'participants': participants
+        }
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -68,10 +116,12 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    if  request.user != room.host:
+        return HttpResponse('you can`t not edit this room')
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -81,9 +131,22 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if  request.user != room.host:
+        return HttpResponse('you can`t not deleted this room')
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html', {'obj': room})
+
+@login_required(login_url='login')
+def deleteMessage(request, pk):
+    messages = Message.objects.get(id=pk)
+    if  request.user != messages.user:
+        return HttpResponse('you can`t not deleted this messages')
+    if request.method == 'POST':
+        messages.delete()
+        return redirect('home')
+    return render(request, 'base/delete.html', {'obj': messages})
